==> 01.Train_test_split.ipynb <==
import numpy as np
import pandas as pd
import os

df = pd.read_excel("/content/drive/MyDrive/Colab Notebooks/Table to insights/Data/Raw Data/table_insights_labeled_data.xlsx",sheet_name = "Data")

output_directory = '/content/drive/MyDrive/Colab Notebooks/Table to insights/Data/Train_test_split'
os.makedirs(output_directory, exist_ok=True)

# Iterate through unique DeviceMaskedIDs
for device_id in df['DeviceMaskedID'].unique():
    device_folder = os.path.join(output_directory, f'{device_id}')
    os.makedirs(device_folder, exist_ok=True)

    # Iterate through train and test data
    for data_type in df['data (train/ test)'].unique():
        # Select data for the current DeviceMaskedID and data type
        filtered_data = df[(df['DeviceMaskedID'] == device_id) & (df['data (train/ test)'] == data_type)][["date","cpu_usage","ram_usage","diskio_usage"]]

        # Save the data to a CSV file
        file_path = os.path.join(device_folder, f'{data_type}.csv')
        filtered_data.to_csv(file_path, index=False)

print("Folders and CSV files created successfully.")


df = pd.read_excel("/content/drive/MyDrive/Colab Notebooks/Table to insights/Data/Raw Data/table_insights_labeled_data.xlsx",sheet_name = "Labels")

# Iterate through unique DeviceMaskedIDs
for device_id in df['DeviceMaskedID'].unique():
    device_folder = os.path.join('/content/drive/MyDrive/Colab Notebooks/Table to insights/Data/Train_test_split', f'{device_id}')
    os.makedirs(device_folder, exist_ok=True)

    # Iterate through data types (train and test)
    for data_type in df['data (train/ test)'].unique():
        # Iterate through columns insight_0, insight_1, insight_2
        # Save the current column to a new txt file
        file_name = f'{data_type}_insights.csv'
        file_path = os.path.join(device_folder, file_name)

        # Create a DataFrame with DeviceMaskedID, data type, and the current column
        filtered_data = df[(df['DeviceMaskedID'] == device_id) & (df['data (train/ test)'] == data_type)][['insight_0', 'insight_1', 'insight_2']]

        # Save the DataFrame to the CSV file
        filtered_data.to_csv(file_path, index=False)


print("Folders and insights txt files created successfully.")
