==> 02.convert to Json.ipynb <==
import numpy as np
import pandas as pd
import os
import json

# Specify the path to the directory
directory_path = '/content/drive/MyDrive/Colab Notebooks/Table to insights/Data/Train_test_split'

# Get all folder names in the directory
folder_names = [folder for folder in os.listdir(directory_path) if os.path.isdir(os.path.join(directory_path, folder))]
print(folder_names)

output_directory = '/content/drive/MyDrive/Colab Notebooks/Table to insights/Data/Processed_Data'
os.makedirs(output_directory, exist_ok=True)
output_directory_train = '/content/drive/MyDrive/Colab Notebooks/Table to insights/Data/Processed_Data/Train_data'
os.makedirs(output_directory, exist_ok=True)
output_directory_test = '/content/drive/MyDrive/Colab Notebooks/Table to insights/Data/Processed_Data/Test_data'
os.makedirs(output_directory, exist_ok=True)
no_train_data = []
no_test_data = []
for curr_folder in folder_names:
    curr_path = os.path.join(directory_path, curr_folder)

    # Process Train and Test data for the current folder
    """
    Data is processed for the current folder, combining Train and Test datasets with their respective insights.
    If data is not available for Train or Test, a message is printed.
    """

    # Process Train data
    train_data = pd.read_csv(os.path.join(curr_path, "Train.csv"))
    if train_data.shape[0] == 0:
      test_data = pd.read_csv(os.path.join(curr_path, "Test.csv"))
      no_train_data.append(f'{curr_folder}')
      test_json = test_data.to_json(orient="records")
      test_insights = pd.read_csv(os.path.join(curr_path, "Test_insights.csv"))
      test_insights_json = test_insights.to_json(orient="records")
      combined_test_data = {"Data": json.loads(test_json), "Insights": json.loads(test_insights_json)}
      device_folder = os.path.join(output_directory_test, f'{curr_folder}')
      os.makedirs(device_folder, exist_ok=True)
      test_file_path = os.path.join(device_folder, "Test.json")
      with open(test_file_path, "w") as json_file:
          json.dump(combined_test_data, json_file, indent=2)
    else:
        no_test_data.append(f'{curr_folder}')
        train_json = train_data.to_json(orient="records")
        train_insights = pd.read_csv(os.path.join(curr_path, "Train_insights.csv"))
        train_insights_json = train_insights.to_json(orient="records")
        combined_train_data = {"Data": json.loads(train_json), "Insights": json.loads(train_insights_json)}
        device_folder = os.path.join(output_directory_train, f'{curr_folder}')
        os.makedirs(device_folder, exist_ok=True)
        train_file_path = os.path.join(device_folder, "Train.json")
        with open(train_file_path, "w") as json_file:
            json.dump(combined_train_data, json_file, indent=2)


print(f"These devices are part of test data set {sorted(no_train_data)}")
print(f"There devices are part of train data set {sorted(no_test_data)}")








print(f"length of test set is {len(no_train_data)} \nlength of train set is {len(no_test_data)}")
