==> 03.Create Hf dataset.ipynb <==
!pip install datasets -q


import os
import json
from datasets import Dataset, DatasetDict
import yaml
from tqdm import tqdm as tqdm_bar

def process_folder(folder_path,filename , prompts):
    dataset = {"table": [], "prompt": [], "label": []}

    for foldername in tqdm_bar(os.listdir(folder_path),desc=filename):
        file_path = os.path.join(folder_path, foldername,filename)

        if os.path.isfile(file_path):
            with open(file_path, 'r') as file:
                file_content = json.load(file)
                # Extract data and insights
                table_data = file_content.get("Data", [])
                insights = file_content.get("Insights", [])

                # Repeat the table for each insight
                for idx,insight in enumerate(insights[0].values()):
                    # Add the entry to the dataset
                    dataset["table"].append(table_data)
                    dataset["prompt"].append(prompts[idx].replace("{table_data}" , str(table_data)))
                    dataset["label"].append(insight)

    return dataset

def create_and_save_dataset(input_folder, output_file , prompts):
    # Process both Train_data and Test_data folders
    train_dataset_dict = process_folder(os.path.join(input_folder, "Train_data"),"Train.json",prompts)
    test_dataset_dict = process_folder(os.path.join(input_folder, "Test_data"),"Test.json" , prompts)

    # Combine train and test datasets
    combined_dataset_dict = {
        "table": train_dataset_dict["table"] + test_dataset_dict["table"],
        "prompt": train_dataset_dict["prompt"] + test_dataset_dict["prompt"],
        "label": train_dataset_dict["label"] + test_dataset_dict["label"],
    }

    # Calculate the index to split the data into train and validation
    split_index = len(train_dataset_dict["table"])

    # Create a Hugging Face Dataset
    combined_dataset = Dataset.from_dict(combined_dataset_dict)

    # Split the combined dataset into train and validation
    dataset_dict = DatasetDict({
        "train": combined_dataset.select(indices=list(range(0, split_index))),
        "validation": combined_dataset.select(indices=list(range(split_index, len(combined_dataset)))),
    })

    # Save the combined dataset to a file
    return dataset_dict

# Get the data dictionary
dict_file_path = "/content/drive/MyDrive/Colab Notebooks/Table to insights/Config/data_dictionary.yaml"

# Read the YAML file
with open(dict_file_path, 'r') as file:
    yaml_content = yaml.safe_load(file)

# Extract the data after the "columns" key
extracted_data = yaml_content.get("columns", [])

# Print the extracted data
for i in extracted_data :
  print(i)

# Add Prompts for each insights to be generated is where the data will be appended
query1 = "What does the data contains and say about. Respond within 2 Lines"
query2 = "What are the top 3 most important insights"
query3 = "What are the top 3 abberations present in the dataset"

prompts = []

prompt_template = "Answer the below query based on the following table , there is a Data Dictionary provided to give context for each column\n \
\n\
Table:\n\
{table_data}\
\n\
Data Dictionary:\n\
{data_dictionary}\n\
\n\
Query:\n\
{query}"

# Update template with each query and append to prompts
prompts.append(prompt_template.format(table_data="{table_data}" ,data_dictionary=extracted_data ,query=query1))
prompts.append(prompt_template.format(table_data="{table_data}" ,data_dictionary=extracted_data ,query=query2))
prompts.append(prompt_template.format(table_data="{table_data}" ,data_dictionary=extracted_data ,query=query3))

# Process both Train_data and Test_data folders
analytical_dataset = create_and_save_dataset("/content/drive/MyDrive/Colab Notebooks/Table to insights/Data/Processed_Data",\
                        "/content/drive/MyDrive/Colab Notebooks/Table to insights/Data/Analytical Datset/Hf_compatible_master_data",
                        prompts)

analytical_dataset

print(f"The example for a prompt is:\n{analytical_dataset['train'][0]['prompt']}\n{'-'*100}\nThe response generated is:\n{analytical_dataset['train'][0]['label']}")

print(f"The example for a prompt is:\n{analytical_dataset['validation'][0]['prompt']}\n{'-'*100}\nThe response generated is:\n{analytical_dataset['validation'][0]['label']}")

analytical_dataset.save_to_disk("/content/drive/MyDrive/Colab Notebooks/Table to insights/Data/Analytical Datset")
